refactor(ui): extract mobile navigation orchestration
